simplify safe/unsafe code files and builds
